requirements.txt
certifi==2024.8.30
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
